refactor: purge the MySQL-era surface that renders invalid or lying SQL
Table options that were Postgres syntax errors, the Year type whose
only render was a panic, money's impossible arguments, the unsigned
types that rendered signed, the duplicate spellings of smallint and
timestamp, the never-constructed JoinOn variant and the dead
DistinctRow — all deleted or collapsed, with Bytea chosen over Blob
because it names the type it renders, and Timestamp taking the
honest tz-naive chrono mapping that finally round-trips. The serial
substitution narrows to serial_spelling over the integer trio — the
full flag redesign and a validation error were both rejected, the
latter because the derive marks every single-column key
auto-increment and a refusal would reject ordinary Uuid keys. The
DDL-text bridge's reverse map adapts and turns near-bijective, its
one surviving collapse stated. Thirteen forward rules and the bridge
rule bump with every reference re-pinned; prepare_column_type is
total and the mysql-options oracle pin becomes deletion-proof. Net
function deletion in query_builder.rs ends the standing density
bypass.

nplan-id: unrep.mysql-purge
nplan-rule: [spec:pgorm:def:sql.ast.select]
nplan-rule: [spec:pgorm:def:sql.render.ddl.types]
nplan-rule: [spec:pgorm:def:sql.types.column-type]
nplan-rule: [spec:pgorm:def:sql.value.value-type]
nplan-rule: [spec:pgorm:req:sql.ddl.column-def]
nplan-rule: [spec:pgorm:req:sql.ddl.column-types]
nplan-rule: [spec:pgorm:req:sql.ddl.create-table]
nplan-rule: [spec:pgorm:req:sql.render.joins]
nplan-rule: [spec:pgorm:sem:codegen.ddl.types]
nplan-rule: [spec:pgorm:sem:codegen.entity.compact.attrs]
nplan-rule: [spec:pgorm:sem:codegen.entity.types.datetime]
nplan-rule: [spec:pgorm:sem:codegen.entity.types]
nplan-rule: [spec:pgorm:sem:macros.derive.entity-model.column-def]
nplan-rule: [spec:pgorm:sem:sql.ddl.panics]
nplan-rule: [spec:pgorm:def:sql.value.value-type/test]
nplan-rule: [spec:pgorm:sem:macros.derive.entity-model.column-def/test]
nplan-rule: [spec:pgorm:sem:sql.ddl.panics/test]

+++ pgorm-codegen/tests/compact/rust_keyword.rs
@@ -5,14 +5,14 @@ use pgorm::entity::prelude::*;
 pub struct Model {
     #[pgorm(primary_key)]
     pub id: i32,
-    pub testing: i8,
-    pub rust: i8,
+    pub testing: i16,
+    pub rust: i16,
     pub keywords: i16,
     pub r#type: i16,
     pub r#typeof: i32,
-    pub crate_: u32,
+    pub crate_: i64,
     pub self_: i64,
-    pub self_id1: u64,
+    pub self_id1: i64,
     pub self_id2: i32,
     pub fruit_id1: i32,
     pub fruit_id2: i32,
